+++ test/integration/default/inspec/default_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe 'bonusbits_mediawiki_nginx::default' do
-  it 'Include CloudWatch Logs Recipe' do
-    have_entry(cron('0 0 * * * yum -y update --security')).with_user('root')
-  end
-end